requirements.txt
flask
python-dotenv
ultralytics
easyocr
opencv-python
flask-cors
torch
